Temperature converter complete

+++ adt/Temperature/src/com/basu/temperature/Converter.java
@@ -2,11 +2,11 @@ package com.basu.temperature;
 
 public class Converter {
 	
-	public static float converCelsiusToFahrenheit(float celsius) {
+	public static double converCelsiusToFahrenheit(double celsius) {
 		return ((celsius - 32) * 5 / 9);
 	}
 	
-	public static float convertFahrenheitToCelsius(float fahrenheit) {
+	public static double convertFahrenheitToCelsius(double fahrenheit) {
 		return ((fahrenheit * 9) / 5) + 32;
 	}
